+++ documentation/Changelog.md
@@ -14,6 +14,7 @@ As of version 0.8.7, *Time Copy* was transformed into a sidebar-based interface.
 
 * **Other Changes**
   * Code cleanup
+  * Module Amag ProTime: Change click-overlay text + warning text
 
 ## V 1.3.17 (Release)